apps/dhs/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pyyaml==6.0.2
jinja2==3.1.4
prometheus-client==0.21.0
python-json-logger==2.0.7
